numeric::operator int64_t: fix MSVC warning

+++ src/tdscpp.h
@@ -1043,11 +1043,11 @@ namespace tds {
 
         constexpr operator int64_t() noexcept {
             if constexpr (N == 0)
-                return neg ? -low_part : low_part;
+                return neg ? -(int64_t)low_part : low_part;
             else {
                 numeric<0> n = *this;
 
-                return neg ? -n.low_part : n.low_part;
+                return neg ? -(int64_t)n.low_part : n.low_part;
             }
         }
